pyproject.toml
[tool.poetry]
name = "jadongsoseol"
version = "0.1.0"
description = ""
authors = ["oooo12-git <jhdmbwy12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.61.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.3"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
ipykernel = "^6.29.5"
langgraph = "^0.2.69"
google = "^3.0.0"
google-api-python-client = "^2.161.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
langchain-anthropic = "^0.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"